Rename Model/Repository -> Model/Table.
