+++ tests/TestCase/ORM/TableRegistryTest.php
@@ -159,22 +159,22 @@ class TableRegistryTest extends TestCase {
  */
 	public function testBuildConvention() {
 		$table = TableRegistry::get('articles');
-		$this->assertInstanceOf('\TestApp\Model\Repository\ArticlesTable', $table);
+		$this->assertInstanceOf('\TestApp\Model\Table\ArticlesTable', $table);
 		$table = TableRegistry::get('Articles');
-		$this->assertInstanceOf('\TestApp\Model\Repository\ArticlesTable', $table);
+		$this->assertInstanceOf('\TestApp\Model\Table\ArticlesTable', $table);
 
 		$table = TableRegistry::get('authors');
-		$this->assertInstanceOf('\TestApp\Model\Repository\AuthorsTable', $table);
+		$this->assertInstanceOf('\TestApp\Model\Table\AuthorsTable', $table);
 		$table = TableRegistry::get('Authors');
-		$this->assertInstanceOf('\TestApp\Model\Repository\AuthorsTable', $table);
+		$this->assertInstanceOf('\TestApp\Model\Table\AuthorsTable', $table);
 
 		$class = $this->getMockClass('\Cake\ORM\Table');
 		$class::staticExpects($this->once())
 			->method('defaultConnectionName')
 			->will($this->returnValue('test'));
 
-		if (!class_exists('MyPlugin\Model\Repository\SuperTestsTable')) {
-			class_alias($class, 'MyPlugin\Model\Repository\SuperTestsTable');
+		if (!class_exists('MyPlugin\Model\Table\SuperTestsTable')) {
+			class_alias($class, 'MyPlugin\Model\Table\SuperTestsTable');
 		}
 
 		$table = TableRegistry::get('MyPlugin.SuperTests');
