+++ tests/test_app/Plugin/TestPlugin/Model/Table/TestPluginCommentsTable.php
@@ -14,15 +14,14 @@
  * @since         CakePHP v 3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
+namespace TestPlugin\Model\Table;
+
+use Cake\ORM\Table;
 
 /**
  * Class TestPluginCommentsTable
  *
  */
-namespace TestPlugin\Model\Repository;
-
-use Cake\ORM\Table;
-
 class TestPluginCommentsTable extends Table {
 
 	protected $_table = 'test_plugin_comments';
