+++ tests/test_app/TestApp/Model/Table/AuthorsTable.php
@@ -9,20 +9,18 @@
  * @since         CakePHP(tm) v 3.0
  * @license       MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace TestApp\Model\Repository;
+namespace TestApp\Model\Table;
 
 use Cake\ORM\Table;
 
 /**
- * Article table class
+ * Author table class
  *
  */
-class ArticlesTable extends Table {
+class AuthorsTable extends Table {
 
 	public function initialize(array $config) {
-		$this->belongsTo('authors');
-		$this->belongsToMany('tags');
-		$this->hasMany('articlesTags');
+		$this->hasMany('articles');
 	}
 
 }
