+++ tests/test_app/TestApp/Model/Table/PostsTable.php
@@ -13,7 +13,7 @@
  * @since         CakePHP v 3.0.0
  * @license       MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace TestApp\Model\Repository;
+namespace TestApp\Model\Table;
 
 use Cake\ORM\Table;
